+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # CHANGELOG MODULE TIMESHEETWEEK FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
+## Non publié
+
+- Ajout des compteurs de zones et de paniers dans l'entête des feuilles hebdomadaires.
+- Affichage du libellé "Zone" devant chaque sélecteur quotidien.
+- Ajout de la traduction "Meals" en "Repas".
+
 ## 1.0
 
 Initial version